Remove streamlit_app.py and upload_model.py
The files 'streamlit_app.py' and 'upload_model.py' have been removed from the project. The former was used for making predictions using the trained model in a Streamlit app and the latter was used for uploading the model to S3. The removal of these files indicates a shift in project structure or execution method.