Adjust training data and time metrics in eval.py
Changed the testing dataset to use CIFAR100 training set, renamed it to 'dataset', and updated shuffling to be True. Also, the inference time calculation is modified to be in milliseconds and adjusted according to the new batch size. These changes are made to enhance the accuracy and performance of the model evaluation.